requirements.txt
Flask==1.1.1
flask-cors
flask-lambda-support==0.1.5
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
marshmallow==3.3.0